refactor(api): migrate route controller
