+++ api/tests/unit/application/organization-learners/index_test.js
@@ -83,25 +83,4 @@ describe('Unit | Application | Router | organization-learner-router', function (
       });
     });
   });
-
-  describe('GET /api/organization-learners', function () {
-    const method = 'GET';
-
-    it('should return a HTTP status code 200', async function () {
-      // given
-      sinon
-        .stub(organizationLearnerController, 'findAssociation')
-        .callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const url = '/api/organization-learners?userId=1&campaignCode=ABCDEF123';
-
-      // when
-      const response = await httpTestServer.request(method, url);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
 });
